requirements.txt
aiobotocore==2.3.4
aiohttp==3.8.1
aioitertools==0.10.0
aiosignal==1.2.0
amqp==5.1.1
asgiref==3.5.2
astroid==2.11.7
asttokens==2.0.5
async-timeout==4.0.2
atomicwrites==1.4.1
attrs==21.4.0
backcall==0.2.0
billiard==3.6.4.0
boto==2.49.0
boto3==1.24.33
botocore==1.27.40
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
dill==0.3.5.1
Django==4.0.4
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-rest-swagger==2.2.0
django-simple-history==3.0.0
django-timezone-field==5.0
djangorestframework==3.13.1
Employee==0.0.4
executing==0.8.3
facebook-sdk==3.1.0
frozenlist==1.3.0
fsspec==2022.5.0
idna==3.3
iniconfig==1.1.1
ipdb==0.13.9
ipython==8.4.0
isort==5.10.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.2.4
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.7.0
mock==4.0.3
multidict==6.0.2
mysqlclient==2.1.1
numpy==1.23.1
openapi-codec==1.3.2
packaging==21.3
pandas==1.4.3
parso==0.8.3
parts==1.5.2
pickleshare==0.7.5
platformdirs==2.5.2
pluggy==1.0.0
products==1.1.2
prompt-toolkit==3.0.30
psycopg2==2.9.3
pure-eval==0.2.2
py==1.11.0
pycryptodome==3.15.0
Pygments==2.12.0
pylint==2.14.5
pyparsing==3.0.9
pytest==7.1.2
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.0
s3fs==2022.5.0
s3transfer==0.6.0
secret==0.8
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
tabulate==0.8.10
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
traitlets==5.3.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
yarl==1.7.2
